基本完了吧

+++ gnote/Gnote/FormMain.cs
@@ -21,11 +21,8 @@ namespace Gnote
             Load_treeview();
             Load_note_list();
         }
-        /**
-         * 左侧导航栏
-         *
-         * */
-         
+        #region 左侧导航栏
+
         //加载左侧导航栏
         public void Load_treeview()
         {
@@ -47,8 +44,8 @@ namespace Gnote
 
             TreeNodeTag treenodefavouriteroottag = new TreeNodeTag();
             TreeNode facourite_root = new TreeNode("收藏夹");
-            facourite_root.Name = "facourite_root";
-            treenodefavouriteroottag.node_type = "facourite_root";
+            facourite_root.Name = "favourite_root";
+            treenodefavouriteroottag.node_type = "favourite_root";
             facourite_root.Tag = treenodefavouriteroottag;
             this.treeview_nav.Nodes.Add(facourite_root);
 
@@ -57,8 +54,9 @@ namespace Gnote
             trash_root.Name = "trash_root";
             treenodetrashroottag.node_type = "trash_root";
             trash_root.Tag = treenodetrashroottag;
+            trash_root.ContextMenuStrip = contextMenuStrip_trash;
             this.treeview_nav.Nodes.Add(trash_root);
- 
+
             //添加笔记本
             Notebook notebook = new Notebook();
             MySqlDataReader reader_notebook = notebook.listNotebook();
@@ -68,23 +66,18 @@ namespace Gnote
                 {
                     //创建新节点对象
                     TreeNode newnode = new TreeNode(reader_notebook.GetString(1));
-                    //创建notebooktag类对象
                     TreeNodeTag treenodenotebooktag = new TreeNodeTag();
-                    //给对象赋值
-                    treenodenotebooktag.notebook_no = reader_notebook.GetInt32(0);
-                    treenodenotebooktag.notebook_name = reader_notebook.GetString(1);
+                    treenodenotebooktag.no = reader_notebook.GetInt32("notebook_no");
+                    treenodenotebooktag.name = reader_notebook.GetString("notebook_name");
                     treenodenotebooktag.is_default = reader_notebook.GetInt32("notebook_default");
                     treenodenotebooktag.node_type = "notebook";
-                    //传给节点
                     newnode.Tag = treenodenotebooktag;
-                    if (treenodenotebooktag.is_default == 1)
-                    {
-                        newnode.BackColor = Color.Silver;
-                    }
-                    
-                    //给节点添加右击菜单
                     newnode.ContextMenuStrip = contextMenuStrip_notebook;
                     //找到父节点
+                    if (treenodenotebooktag.is_default == 1)
+                    {
+                        newnode.ForeColor = Color.Red;
+                    }
                     TreeNode tnParent = treeview_nav.Nodes["notebook_root"];
                     //添加到父节点
                     tnParent.Nodes.Add(newnode);
@@ -99,26 +92,29 @@ namespace Gnote
                 if (reader_label.HasRows)
                 {
                     TreeNode newnode = new TreeNode(reader_label.GetString(1));
+                    TreeNodeTag treenodelabeltag = new TreeNodeTag();
+                    treenodelabeltag.no = reader_label.GetInt32("label_no");
+                    treenodelabeltag.name = reader_label.GetString("label_name");
+                    treenodelabeltag.node_type = "label";
+                    newnode.Tag = treenodelabeltag;
+                    newnode.ContextMenuStrip = contextMenuStrip_label;
                     TreeNode tnParent = treeview_nav.Nodes["label_root"];
                     tnParent.Nodes.Add(newnode);
                 }
             }
+            treeview_nav.ExpandAll();
         }
 
+
+
+        /////////////////////////////右击左侧导航栏//////////////////////////
         //点击刷新左侧导航栏
         private void 刷新ToolStripMenuItem_Click(object sender, EventArgs e)
         {
             treeview_nav.Nodes.Clear();
             this.Load_treeview();
         }
-        // 新建笔记本
-        private void NewNotebook_Click(object sender, EventArgs e)
-        {
-            NewNotebook NewNotebook = new NewNotebook();
-            NewNotebook.Show();
-        }
-
-        //笔记本右击菜单
+        ///////////////////////////笔记本右击菜单///////////////////////////////
         //新建笔记本
         private void 新建笔记本ToolStripMenuItem_Click(object sender, EventArgs e)
         {
@@ -126,9 +122,8 @@ namespace Gnote
             NewNotebook.Show();
         }
 
-        //笔记本管理器
 
-        //笔记右击菜单
+        ///////////////////////////笔记本右击菜单&左键选中/////////////////////////////////
         private void treeview_nav_MouseDown(object sender, MouseEventArgs e)
         {
             if (e.Button == MouseButtons.Right)
@@ -140,33 +135,66 @@ namespace Gnote
             {
                 //左击显示列表
                 treeview_nav.SelectedNode = treeview_nav.GetNodeAt(e.X, e.Y);
-                TreeNodeTag treenodenotebooktag = treeview_nav.SelectedNode.Tag as TreeNodeTag;
-                if (treenodenotebooktag.node_type == "notebook")
+                if (treeview_nav.SelectedNode != null)
                 {
-                    int nodeno = treenodenotebooktag.notebook_no;
-                    Load_note_list(nodeno);
+                    TreeNodeTag treenodenotebooktag = treeview_nav.SelectedNode.Tag as TreeNodeTag;
+                    TreeNodeTag treenodelabeltag = treeview_nav.SelectedNode.Tag as TreeNodeTag;
+                    if (treenodenotebooktag != null || treenodelabeltag != null)
+                    {
+                        if (treenodelabeltag.node_type == "label")
+                        {
+                            int nodeLabelNo = treenodelabeltag.no;
+                            Load_label_list(nodeLabelNo);
+                        }
+                        else if (treenodenotebooktag.node_type == "notebook")
+                        {
+                            int nodeNotebookNo = treenodenotebooktag.no;
+                            Load_note_list(nodeNotebookNo);
+                        }
+                        else if (treenodenotebooktag.node_type == "favourite_root")
+                        {
+                            int nodeNotebookNo = treenodenotebooktag.no;
+                            Load_favourite_list();
+                        }
+                        else if (treenodenotebooktag.node_type == "trash_root")
+                        {
+                            int nodeNotebookNo = treenodenotebooktag.no;
+                            Load_trash_list();
+                        }
+
+                    }
                 }
+                
             }
         }
-        //删除笔记笔记本
+        //删除笔记本
         private void 删除ToolStripMenuItem_Click(object sender, EventArgs e)
         {
-            //创建tag对象
-            TreeNodeTag treenodenotebooktag = treeview_nav.SelectedNode.Tag as TreeNodeTag;
-            //取值
-            string nodename = treenodenotebooktag.notebook_name;
-            int nodeno = treenodenotebooktag.notebook_no;
-            //用控制器删掉
-            Controller.Notebook notebookno = new Controller.Notebook();
-            notebookno.deleteNotebook(nodeno);
-            //刷新
-            treeview_nav.Nodes.Clear();
-            this.Load_treeview();
+            MessageBoxButtons messButton = MessageBoxButtons.OKCancel;
+            DialogResult dr = MessageBox.Show("删除笔记本将会同时删除笔记本内所有笔记，确认删除?", "删除笔记本", messButton);
+            if (dr == DialogResult.OK)//如果点击“确定”按钮
+            {
+                //创建tag对象
+                TreeNodeTag treenodenotebooktag = treeview_nav.SelectedNode.Tag as TreeNodeTag;
+                //取值
+                string nodename = treenodenotebooktag.name;
+                int nodeno = treenodenotebooktag.no;
+                //用控制器删掉
+                Controller.Notebook notebookno = new Controller.Notebook();
+                notebookno.deleteNotebook(nodeno);
+                //刷新
+                treeview_nav.Nodes.Clear();
+                this.Load_treeview();
+            }
+            else//如果点击“取消”按钮
+            {
+
+            }
         }
         private void 设置为默认笔记本ToolStripMenuItem_Click(object sender, EventArgs e)
         {
             TreeNodeTag treenodenotebooktag = treeview_nav.SelectedNode.Tag as TreeNodeTag;
-            int nodeno = treenodenotebooktag.notebook_no;
+            int nodeno = treenodenotebooktag.no;
             Controller.Notebook notebookno = new Controller.Notebook();
             notebookno.setDefault(nodeno);
             treeview_nav.Nodes.Clear();
@@ -174,33 +202,79 @@ namespace Gnote
         }
 
         //重命名笔记本
-        private void ToolStripMenuItem_notebook_Click(object sender, EventArgs e)
+        private void 重命名ToolStripMenuItem_notebook_Click(object sender, EventArgs e)
         {
             treeview_nav.LabelEdit = true;
             treeview_nav.SelectedNode.BeginEdit();
         }
         private void treeview_nav_AfterLabelEdit(object sender, NodeLabelEditEventArgs e)
         {
-            string newTxt = e.Label;//获取新文本
-            //string id = e.Node.Text;//获取原来的文本
-            TreeNodeTag treenodenotebooktag = treeview_nav.SelectedNode.Tag as TreeNodeTag;
-            string nodename = treenodenotebooktag.notebook_name;
-            int nodeno = treenodenotebooktag.notebook_no;
-            Controller.Notebook notebookno = new Controller.Notebook();
-            notebookno.renameNotebook(nodeno, newTxt);
-            treeview_nav.Nodes.Clear();
-            this.Load_treeview();
-            treeview_nav.LabelEdit = false;
+            string newtext = e.Label;//获取新文本
+            string oldtext = e.Node.Text;//获取原来的文本
+            newtext = oldtext;
+            if (newtext != oldtext)
+            {
+                Console.WriteLine("不一样");
+                TreeNodeTag treenodenotebooktag = treeview_nav.SelectedNode.Tag as TreeNodeTag;
+                //获得节点信息
+                string nodename = treenodenotebooktag.name;
+                int nodeno = treenodenotebooktag.no;
+                //判断类型
+                string nodetype = treenodenotebooktag.node_type;
+                if (nodetype == "notebook")
+                {
+                    Controller.Notebook notebookno = new Controller.Notebook();
+                    notebookno.renameNotebook(nodeno, newtext);
+
+                }
+                else if (nodetype == "label")
+                {
+                    Controller.Label label = new Controller.Label();
+                    label.renameLabel(nodeno, newtext);
+                }
+                //刷新
+                treeview_nav.Nodes.Clear();
+                this.Load_treeview();
+                treeview_nav.LabelEdit = false;
+            }
+            else
+            {
+                Console.WriteLine("一样的");
+                treeview_nav.LabelEdit = false;
+            }
+            
         }
 
+        //////////////////////////////标签右击菜单///////////////////////////////
+        //重命名标签
+        private void 重命名ToolStripMenuItem1_Click(object sender, EventArgs e)
+        {
+            treeview_nav.LabelEdit = true;
+            treeview_nav.SelectedNode.BeginEdit();
+        }
 
-        /**
-         * 中间列表栏
-         * 
-         * 
-         * 
-         * */
-        //加载笔记列表
+        //删除标签
+        private void 删除ToolStripMenuItem2_Click(object sender, EventArgs e)
+        {
+            //创建tag对象
+            TreeNodeTag treenodelabeltag = treeview_nav.SelectedNode.Tag as TreeNodeTag;
+            //取值
+            string nodename = treenodelabeltag.name;
+            int nodeno = treenodelabeltag.no;
+            //用控制器删掉
+            Controller.Label label = new Controller.Label();
+            label.deleteLabel(nodeno);
+            //刷新
+            treeview_nav.Nodes.Clear();
+            this.Load_treeview();
+        }
+
+        #endregion
+
+
+        #region 中间列表栏
+
+        //加载所有笔记列表
         public void Load_note_list()
         {
             listBox.Items.Clear();
@@ -216,21 +290,80 @@ namespace Gnote
                 }
             }
         }
-
+        //加载指定笔记本
         public void Load_note_list(int notebook_no)
         {
             listBox.Items.Clear();
             Note note = new Note();
-            MySqlDataReader reader_note = note.listNote(notebook_no);
+            MySqlDataReader reader_note = note.listNote(notebook_no, "notebook");
             while (reader_note.Read())
             {
                 if (reader_note.HasRows)
                 {
-                    this.listBox.Items.Add(reader_note.GetString("note_name"));
+                    string a;
+                    if (reader_note.GetInt32("note_favourite") == 1)
+                    {
+                        a = reader_note.GetString("note_name") + "(收藏)";
+                    }
+                    else
+                    {
+                        a = reader_note.GetString("note_name");
+                    }
+                    listBox.Items.Add(new ListBoxItems { Name = a, Value = reader_note.GetInt32("note_no") });
+                    listBox.ContextMenuStrip = contextMenuStripListBox;
+                    listBox.DisplayMember = "Name";
+                    
                 }
             }
         }
-
+        //加载指定标签笔记
+        public void Load_label_list(int label_no)
+        {
+            listBox.Items.Clear();
+            Note note = new Note();
+            MySqlDataReader reader_note = note.listNote(label_no, "label");
+            while (reader_note.Read())
+            {
+                if (reader_note.HasRows)
+                {
+                    listBox.Items.Add(new ListBoxItems { Name = reader_note.GetString("note_name"), Value = reader_note.GetInt32("note_no") });
+                    listBox.ContextMenuStrip = contextMenuStripListBox;
+                    listBox.DisplayMember = "Name";
+                }
+            }
+        }
+        //加载删除掉的笔记
+        public void Load_trash_list()
+        {
+            listBox.Items.Clear();
+            Note note = new Note();
+            MySqlDataReader reader_note = note.listNote("trash");
+            while (reader_note.Read())
+            {
+                if (reader_note.HasRows)
+                {
+                    listBox.Items.Add(new ListBoxItems { Name = reader_note.GetString("note_name"), Value = reader_note.GetInt32("note_no") });
+                    listBox.ContextMenuStrip = contextMenuStripListBoxTrash; ;
+                    listBox.DisplayMember = "Name";
+                }
+            }
+        }
+        //加载收藏的笔记
+        public void Load_favourite_list()
+        {
+            listBox.Items.Clear();
+            Note note = new Note();
+            MySqlDataReader reader_note = note.listNote("favourite");
+            while (reader_note.Read())
+            {
+                if (reader_note.HasRows)
+                {
+                    listBox.Items.Add(new ListBoxItems { Name = reader_note.GetString("note_name"), Value = reader_note.GetInt32("note_no") });
+                    listBox.ContextMenuStrip = contextMenuStripListBox;
+                    listBox.DisplayMember = "Name";
+                }
+            }
+        }
         //点击笔记
         public void listBox_MouseDown(object sender, MouseEventArgs e)
         {
@@ -244,16 +377,19 @@ namespace Gnote
                 //左击显示列表,显示笔记详情
                 ListBoxItems listboxitems = new ListBoxItems();
                 listboxitems = this.listBox.SelectedItem as ListBoxItems;
-                MySqlDataReader reader_note_detail = new Note().ShowDetail(listboxitems.Value);
-                while (reader_note_detail.Read())
+                if (listboxitems != null)
                 {
-                    if (reader_note_detail.HasRows)
+                    MySqlDataReader reader_note_detail = new Note().ShowDetail(listboxitems.Value);
+                    while (reader_note_detail.Read())
                     {
-                        textBoxTitle.Text = reader_note_detail.GetString("note_name");
-                        richTextBoxEdit.Text = reader_note_detail.GetString("note_content");
+                        if (reader_note_detail.HasRows)
+                        {
+                            textBoxTitle.Text = reader_note_detail.GetString("note_name");
+                            textBoxEdit.Text = reader_note_detail.GetString("note_content");
+                        }
                     }
                 }
-                
+
             }
         }
         // 点击搜索按钮
@@ -262,45 +398,98 @@ namespace Gnote
 
         }
 
-        //测试按钮
-        private void test_Click(object sender, EventArgs e)
-        {
-            Notebook notebook = new Notebook();
-            notebook.listNotebook();
-        }
+        #endregion  
 
-        /**
-         * 菜单栏
-         * 
-         * */
+
+        #region 工具栏
+        // 新建笔记本
+        private void NewNotebook_Click(object sender, EventArgs e)
+        {
+            NewNotebook NewNotebook = new NewNotebook();
+            NewNotebook.Show();
+            Load_treeview();
+        }
+        #endregion
+
+
+
+        #region 菜单栏
+        //////////////文件
         //新建笔记
         private void 新建ToolStripMenuItem_Click(object sender, EventArgs e)
         {
-            Console.WriteLine("点击了新建笔记");
             textBoxTitle.Text = "无标题笔记";
-            richTextBoxEdit.Text = "";
+            textBoxEdit.Text = "";
+            //添加到默认笔记本
             new Note().newNote(new Notebook().getDefault());
-            //重新加载笔记列表
+            //显示默认笔记本列表
+            Load_note_list(new Notebook().getDefault());
+            toolStripStatusLabelTip.Text = "笔记新建成功";
+            
+        }
+        //保存
+        private void 保存ToolStripMenuItem_Click(object sender, EventArgs e)
+        {
+            ListBoxItems listboxitems = listBox.SelectedItem as ListBoxItems;
+            if (listboxitems != null)
+            {
+                new Note().SaveNote(listboxitems.Value, textBoxTitle.Text, DateTime.Now, textBoxEdit.Text);
+                Load_note_list();
+                toolStripStatusLabelTip.Text = "保存成功";
+            }
+        }
+        ////////////////////////////////右击笔记///////////////////////////////////
+        //删除笔记
+        private void 删除ToolStripMenuItem1_Click(object sender, EventArgs e)
+        {
+            ListBoxItems listboxitems = listBox.SelectedItem as ListBoxItems;
+            new Note().deleteNote(listboxitems.Value);
             Load_note_list();
+            toolStripStatusLabelTip.Text = "删除成功";
         }
 
 
+        #endregion
 
 
-        /**
-         * 状态栏
-         * 
-         * */
+        #region 状态栏
 
-        
+        #endregion
+
+
+        #region 右侧编辑栏
+        //加粗
+        private void toolStripButtonBold_Click(object sender, EventArgs e)
+        {
+            Font oldFont = textBoxEdit.SelectionFont;
+            Font newFont;
+            if (oldFont.Bold)
+            {
+                newFont = new Font(oldFont, oldFont.Style & ~FontStyle.Bold);
+            }
+            else
+            {
+                newFont = new Font(oldFont, oldFont.Style | FontStyle.Bold);
+            }
+            textBoxEdit.SelectionFont = newFont;
+        }
+        #endregion
+
+
+
+        public string a { get; set; }
     }
+    #region treeview节点类
     public class TreeNodeTag
     {
-        public int notebook_no;
-        public string notebook_name;
+        public int no;
+        public string name;
         public int is_default;
         public string node_type;
     }
+    #endregion
+
+    #region listboxitem类
     public class ListBoxItems
     {
         private string _name;
@@ -319,4 +508,5 @@ namespace Gnote
         }
 
     }
+    #endregion
 }
